format

+++ src/main/java/com/cibertec/dsw2/Repository/CardRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface CardRepository extends JpaRepository<Card,Long> {
-
+public interface CardRepository extends JpaRepository<Card,Integer> {
 }
